build: switch to gradle
