+++ gradle/subproject-release.gradle
@@ -1,0 +1,78 @@
+apply plugin: 'signing'
+
+signing {
+    required {
+        isReleaseVersion
+    }
+
+    sign configurations.archives
+}
+
+uploadArchives {
+    repositories {
+        mavenDeployer {
+            repository(url: project.nexusUrl) {
+                authentication(
+                        userName: project.nexusUsername,
+                        password: project.nexusPassword
+                )
+            }
+
+            beforeDeployment {
+                MavenDeployment deployment -> signing.signPom(deployment)
+            }
+
+            pom.groupId = project.group
+            pom.artifactId = project.name
+            pom.version = project.version
+
+            pom.project {
+                name project.name
+                description project.name
+                url project.gitProject
+
+                scm {
+                    url project.gitProject
+                    connection "scm:${project.gitProject}"
+                    developerConnection "scm:${project.gitURL}"
+                }
+
+                licenses {
+                    license {
+                        name 'The Apache Software License, Version 2.0'
+                        url 'http://www.apache.org/licenses/LICENSE-2.0.txt'
+                        distribution 'repo'
+                    }
+                }
+
+                developers {
+                    developer {
+                        name 'Fanmin Shi'
+                        email 'fanmin.shi@coreos.com'
+                        organization 'coreos'
+                        organizationUrl 'http://coreos.com'
+                    }
+
+                    developer {
+                        name 'Xiang Li'
+                        email 'xiang.li@coreos.com'
+                        organization 'coreos'
+                        organizationUrl 'http://coreos.com'
+                    }
+
+                    developer {
+                        name 'Anthony Romano'
+                        email 'anthony.romano@coreos.com<'
+                        organization 'coreos'
+                        organizationUrl 'http://coreos.com'
+                    }
+
+                    developer {
+                        name 'Luca Burgazzoli'
+                        email 'lburgazzoli@gmail.com'
+                    }
+                }
+            }
+        }
+    }
+}
