--- jetcd-all/pom.xml
@@ -1,141 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Copyright 2017 The jetcd authors
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-        http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-
--->
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-
-    <parent>
-        <groupId>com.coreos</groupId>
-        <artifactId>jetcd-parent</artifactId>
-        <version>0.1.0-SNAPSHOT</version>
-    </parent>
-
-    <modelVersion>4.0.0</modelVersion>
-    <artifactId>jetcd-all</artifactId>
-    <packaging>bundle</packaging>
-
-    <dependencies>
-        <dependency>
-            <groupId>com.coreos</groupId>
-            <artifactId>jetcd-common</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.coreos</groupId>
-            <artifactId>jetcd-core</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.coreos</groupId>
-            <artifactId>jetcd-resolver</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.coreos</groupId>
-            <artifactId>jetcd-resolver-dns-srv</artifactId>
-        </dependency>
-    </dependencies>
-
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-shade-plugin</artifactId>
-                <executions>
-                    <execution>
-                        <phase>package</phase>
-                        <goals>
-                            <goal>shade</goal>
-                        </goals>
-                        <configuration>
-                            <createSourcesJar>true</createSourcesJar>
-                            <createDependencyReducedPom>false</createDependencyReducedPom>
-                            <shadedArtifactAttached>false</shadedArtifactAttached>
-                            <minimizeJar>false</minimizeJar>
-                            <artifactSet>
-                                <excludes>
-                                    <exclude>org.slf4j:slf4j-api</exclude>
-                                    <exclude>com.google.errorprone:error_prone_annotations</exclude>
-                                </excludes>
-                            </artifactSet>
-                            <relocations>
-                                <relocation>
-                                    <pattern>com.google</pattern>
-                                    <shadedPattern>com.coreos.jetcd.shaded.com.google
-                                    </shadedPattern>
-                                </relocation>
-                                <relocation>
-                                    <pattern>io.grpc</pattern>
-                                    <shadedPattern>com.coreos.jetcd.shaded.io.grpc</shadedPattern>
-                                </relocation>
-                                <relocation>
-                                    <pattern>io.netty</pattern>
-                                    <shadedPattern>com.coreos.jetcd.shaded.io.netty</shadedPattern>
-                                </relocation>
-                                <relocation>
-                                    <pattern>javax.annotation</pattern>
-                                    <shadedPattern>com.coreos.jetcd.shaded.javax.annotation</shadedPattern>
-                                </relocation>
-                                <relocation>
-                                    <pattern>io.opencensus</pattern>
-                                    <shadedPattern>com.coreos.jetcd.shaded.io.opencensus
-                                    </shadedPattern>
-                                </relocation>
-                            </relocations>
-                        </configuration>
-                    </execution>
-                </executions>
-            </plugin>
-
-            <plugin>
-                <groupId>org.apache.felix</groupId>
-                <artifactId>maven-bundle-plugin</artifactId>
-                <extensions>true</extensions>
-                <configuration>
-                    <instructions>
-                        <Bundle-SymbolicName>${project.groupId}.${project.artifactId}</Bundle-SymbolicName>
-                        <Bundle-Name>CoreOS :: ${project.artifactId}</Bundle-Name>
-                        <Bundle-Version>${project.version}</Bundle-Version>
-                        <Export-Package>com.coreos.jetcd.*</Export-Package>
-                        <Export-Package>
-                            com.coreos.jetcd.common.exception,
-                            com.coreos.jetcd.resolver,
-                            com.coreos.jetcd.resolver.dnssrv,
-                            com.coreos.jetcd,
-                            com.coreos.jetcd.auth,
-                            com.coreos.jetcd.cluster,
-                            com.coreos.jetcd.data,
-                            com.coreos.jetcd.kv,
-                            com.coreos.jetcd.lease,
-                            com.coreos.jetcd.lock,
-                            com.coreos.jetcd.maintenance,
-                            com.coreos.jetcd.op,
-                            com.coreos.jetcd.options,
-                            com.coreos.jetcd.watch
-                        </Export-Package>
-                        <Import-Package>
-                            org.slf4j;version="[1.7,2)",
-                            javax.net.ssl;resolution:=optional
-                        </Import-Package>
-                    </instructions>
-                    <unpackBundle>true</unpackBundle>
-                </configuration>
-            </plugin>
-
-        </plugins>
-    </build>
-
-</project>
