+++ jetcd-all/src/main/java/com/coreos/jetcd/all/JetcdAll.java
@@ -13,9 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package com.coreos.jetcd.all;
 
 public final class JetcdAll {    
-    private JetcdAll() {        
-    }
+  private JetcdAll() {
+  }
 }
