+++ jetcd-examples/jetcd-simple-ctl/build.gradle
@@ -1,0 +1,14 @@
+apply plugin: 'java'
+apply plugin: 'application'
+
+dependencies {
+    compile project(':jetcd-core')
+
+    compile "com.beust:jcommander:1.72"
+
+    runtime "org.apache.logging.log4j:log4j-api:$versions.log4j2"
+    runtime "org.apache.logging.log4j:log4j-core:$versions.log4j2"
+    runtime "org.apache.logging.log4j:log4j-slf4j-impl:$versions.log4j2"
+}
+
+mainClassName = 'com.coreos.jetcd.examples.jetcdctl.Main'
