+++ jetcd-karaf/build.gradle
@@ -1,0 +1,63 @@
+apply from: "$rootDir/gradle/subproject-release.gradle"
+apply plugin: "com.github.lburgazzoli.karaf"
+
+configurations {
+    karaf
+}
+
+compileJava.dependsOn tasks.getByPath(':jetcd-all:install')
+compileJava.dependsOn tasks.getByPath(':jetcd-osgi:install')
+
+dependencies {
+    testCompile (group: "junit", name: "junit", version: versions.junit)
+    testCompile (group: "org.assertj", name: "assertj-core", version: versions.assertj)
+    testCompile (group: "org.ops4j.pax.exam", name: "pax-exam-container-karaf", version: versions.pax_exam)
+    testCompile (group: "org.ops4j.pax.exam", name: "pax-exam-junit4", version: versions.pax_exam)
+    testCompile (group: "org.apache.geronimo.specs", name: "geronimo-atinject_1.0_spec", version: 1.0)
+
+    testCompile (group: "org.apache.karaf.itests", name: "itests", version: versions.karaf, classifier: "tests")
+    testCompile (group: "org.apache.karaf", name: "apache-karaf-minimal", version: versions.karaf, ext: "tar.gz") {
+        exclude group: 'org.apache.karaf', module: 'org.apache.karaf.client'
+    }
+
+    // test runtime
+    testCompile project(":jetcd-core")
+    testCompile project(":jetcd-resolver")
+
+    testRuntime "org.apache.logging.log4j:log4j-api:$versions.log4j2"
+    testRuntime "org.apache.logging.log4j:log4j-core:$versions.log4j2"
+    testRuntime "org.apache.logging.log4j:log4j-slf4j-impl:$versions.log4j2"
+    testRuntime "org.apache.logging.log4j:log4j-1.2-api:$versions.log4j2"
+
+    karaf("com.coreos:jetcd-all:$versions.jetcd") {
+        transitive = false
+    }
+    karaf("com.coreos:jetcd-osgi:$versions.jetcd") {
+        transitive = false
+    }
+}
+
+test {
+    versions.each {
+        k, v -> systemProperty("version.$k", v)
+
+    }
+
+    systemProperty "build.dir", project.buildDir
+}
+
+karaf {
+    features {
+        xsdVersion  = '1.3.0'
+        name        = 'coreos-jetcd'
+        version     = project.version
+        description = 'coreos-jetcd'
+
+        feature {
+            name            = 'coreos-jetcd'
+            description     = 'coreos-jetcd'
+            version         = project.version
+            configurations 'karaf'
+        }
+    }
+}
