--- jetcd-osgi/jetcd-karaf/src/main/resources/features.xml
@@ -1,89 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<!--
-
-    Copyright 2017 The jetcd authors
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-        http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-
--->
-<features xmlns="http://karaf.apache.org/xmlns/features/v1.4.0" name="coreos-jetcd">
-    <feature name="coreos-jetcd" description="coreos-jetcd" version="${project.version}}">
-
-        <details>CoreOS :: Java Client for etcd V3</details>
-
-        <feature prerequisite="true">wrap</feature>
-
-        <!-- google -->
-        <!--
-        <bundle dependency="true">mvn:com.google.code.gson/gson/${gson.version}</bundle>
-        <bundle dependency="true">mvn:com.google.guava/guava/${guava.version}</bundle>
-        <bundle dependency="true">mvn:com.google.protobuf/protobuf-java/${protobuf.version}</bundle>
-        <bundle dependency="true">mvn:com.google.protobuf/protobuf-java-util/${protobuf.version}</bundle>
-        <bundle dependency="true">wrap:mvn:com.google.api.grpc/proto-google-common-protos/${google.grpc.common.version}$Bundle-SymbolicName=proto-google-common-protos&amp;Bundle-Version=${google.grpc.common.version}</bundle>
-        -->
-
-        <!-- misc -->
-        <!--
-        <bundle dependency="true">wrap:mvn:io.opencensus/opencensus-api/${opencensus.version}$Bundle-SymbolicName=opencensus-api&amp;Bundle-Version=${opencensus.version}</bundle>
-        <bundle dependency="true">wrap:mvn:io.opencensus/opencensus-contrib-grpc-metrics/${opencensus.version}$Bundle-SymbolicName=opencensus-contrib-grpc-metrics&amp;Bundle-Version=${opencensus.version}</bundle>
-        -->
-
-        <!-- netty -->
-        <!--
-        <bundle dependency="true">mvn:io.netty/netty-codec-http2/${netty.version}</bundle>
-        <bundle dependency="true">mvn:io.netty/netty-codec-http/${netty.version}</bundle>
-        <bundle dependency="true">mvn:io.netty/netty-codec/${netty.version}</bundle>
-        <bundle dependency="true">mvn:io.netty/netty-handler/${netty.version}</bundle>
-        <bundle dependency="true">mvn:io.netty/netty-buffer/${netty.version}</bundle>
-        <bundle dependency="true">mvn:io.netty/netty-common/${netty.version}</bundle>
-        <bundle dependency="true">mvn:io.netty/netty-handler-proxy/${netty.version}</bundle>
-        <bundle dependency="true">mvn:io.netty/netty-transport/${netty.version}</bundle>
-        <bundle dependency="true">mvn:io.netty/netty-resolver/${netty.version}</bundle>
-        <bundle dependency="true">mvn:io.netty/netty-codec-socks/${netty.version}</bundle>
-        -->
-
-        <!--
-            gRPC
-
-            gRPC has a split package issues that causes troubles with OSGi and Java 9
-            module system, as reference see:
-
-              https://github.com/grpc/grpc-java/issues/3522
-              https://github.com/grpc/grpc-java/issues/2847
-
-            Because of that it is not yet possible to provide a full working solution
-            for OSGi
-         -->
-        <!--
-        <bundle dependency="true">wrap:mvn:io.grpc/grpc-core/${grpc.bundle.version}$Bundle-SymbolicName=grpc-core&amp;Bundle-Version=${grpc.bundle.version}</bundle>
-        <bundle dependency="true">wrap:mvn:io.grpc/grpc-context/${grpc.bundle.version}$Bundle-SymbolicName=grpc-context&amp;Bundle-Version=${grpc.bundle.version}</bundle>
-        <bundle dependency="true">wrap:mvn:io.grpc/grpc-netty/${grpc.bundle.version}$Bundle-SymbolicName=grpc-netty&amp;Bundle-Version=${grpc.bundle.version}</bundle>
-        <bundle dependency="true">wrap:mvn:io.grpc/grpc-protobuf/${grpc.bundle.version}$Bundle-SymbolicName=grpc-protobuf&amp;Bundle-Version=${grpc.bundle.version}</bundle>
-        <bundle dependency="true">wrap:mvn:io.grpc/grpc-protobuf-lite/${grpc.bundle.version}$Bundle-SymbolicName=grpc-protobuf-lite&amp;Bundle-Version=${grpc.bundle.version}</bundle>
-        <bundle dependency="true">wrap:mvn:io.grpc/grpc-stub/${grpc.bundle.version}$Bundle-SymbolicName=grpc-stub&amp;Bundle-Version=${grpc.bundle.version}</bundle>
-        -->
-
-        <!-- jetcd -->
-        <!--
-        <bundle>mvn:com.coreos/jetcd-common/${project.version}</bundle>
-        <bundle>mvn:com.coreos/jetcd-resolver/${project.version}</bundle>
-        <bundle>mvn:com.coreos/jetcd-core/${project.version}</bundle>
-        <bundle>mvn:com.coreos/jetcd-resolver-dns-srv/${project.version}</bundle>
-        -->
-
-        <!-- use jetcd-all until grpc fixes split package -->
-        <bundle>mvn:com.coreos/jetcd-all/${project.version}</bundle>
-        <bundle>mvn:com.coreos/jetcd-osgi/${project.version}</bundle>
-
-    </feature>
-</features>
