+++ jetcd-resolver-dns-srv/build.gradle
@@ -1,0 +1,8 @@
+apply from: "$rootDir/gradle/subproject.gradle"
+apply from: "$rootDir/gradle/subproject-osgi.gradle"
+apply from: "$rootDir/gradle/subproject-release.gradle"
+
+dependencies {
+    compile project(':jetcd-common')
+    compile project(':jetcd-resolver')
+}
