+++ jetcd-resolver/build.gradle
@@ -1,0 +1,7 @@
+apply from: "$rootDir/gradle/subproject.gradle"
+apply from: "$rootDir/gradle/subproject-osgi.gradle"
+apply from: "$rootDir/gradle/subproject-release.gradle"
+
+dependencies {
+    compile project(':jetcd-common')
+}
